Completed chapter 7.
Getting the following error after saving a product edit.
Warning: Can't perform a React state update on an unmounted component. This is a no-op, but it
indicates a memory leak in your application. To fix, cancel all subscriptions
and asynchronous tasks in a useEffect cleanup function

Might have something to do with needing to use useMutation inside our
functional component